Criando funções generalistas para simplificar geração de items na tela Refatorando
