+++ lib/pages/dashBoard/views/dashBoardBody.dart
@@ -1,17 +1,14 @@
 import 'package:flutter/cupertino.dart';
 
-import '../../dashboardAulas/viewAulas.dart';
+import '../../dashboardAulas/dashboard_view.dart';
 
 class DashBoardBody extends StatelessWidget {
   const DashBoardBody({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-       children: [
-         Expanded(child: ViewAulas()),
-       ]
-
-    );
+    return Column(children: [
+      Expanded(child: Text("")),
+    ]);
   }
 }
