+++ lib/pages/dashboardAulas/views/youtube_player.dart
@@ -1,0 +1,129 @@
+import 'dart:developer';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:youtube_player_iframe/youtube_player_iframe.dart';
+
+class YoutubePlayer extends StatefulWidget {
+  List<String> ids = [];
+
+  YoutubePlayer(this.ids, {Key? key}) : super(key: key);
+
+  @override
+  _YoutubePlayer createState() => _YoutubePlayer(ids);
+}
+
+class _YoutubePlayer extends State<YoutubePlayer> {
+  late YoutubePlayerController _controller;
+
+  List<String> ids = [];
+
+  _YoutubePlayer(this.ids);
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = YoutubePlayerController(
+      initialVideoId: ids[0],
+      params: YoutubePlayerParams(
+        playlist: ids,
+        showControls: true,
+        showFullscreenButton: true,
+        desktopMode: false,
+        privacyEnhanced: true,
+        useHybridComposition: true,
+      ),
+    );
+    _controller.onEnterFullscreen = () {
+      SystemChrome.setPreferredOrientations([
+        DeviceOrientation.landscapeLeft,
+        DeviceOrientation.landscapeRight,
+      ]);
+      log('Entered Fullscreen');
+    };
+    _controller.onExitFullscreen = () {
+      log('Exited Fullscreen');
+    };
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    const player = YoutubePlayerIFrame();
+    return YoutubePlayerControllerProvider(
+      // Passing controller to widgets below.
+      controller: _controller,
+
+      child: Scaffold(
+        body: LayoutBuilder(
+          builder: (context, constraints) {
+            if (kIsWeb && constraints.maxWidth > 800) {
+              return Row(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Expanded(child: player),
+                  Column(
+                    children: [
+                      SizedBox(
+                        width: 300,
+                        child: SingleChildScrollView(),
+                      ),
+                    ],
+                  ),
+                ],
+              );
+            }
+            return ListView(
+              children: [
+                Stack(
+                  children: [
+                    player,
+                    Positioned.fill(
+                      child: YoutubeValueBuilder(
+                        controller: _controller,
+                        builder: (context, value) {
+                          return AnimatedCrossFade(
+                            firstChild: const SizedBox.shrink(),
+                            secondChild: Material(
+                              child: DecoratedBox(
+                                decoration: BoxDecoration(
+                                  image: DecorationImage(
+                                    image: NetworkImage(
+                                      YoutubePlayerController.getThumbnail(
+                                        videoId:
+                                            _controller.params.playlist.first,
+                                        quality: ThumbnailQuality.high,
+                                      ),
+                                    ),
+                                    fit: BoxFit.fitWidth,
+                                  ),
+                                ),
+                                child: const Center(
+                                  child: CircularProgressIndicator(),
+                                ),
+                              ),
+                            ),
+                            crossFadeState: value.isReady
+                                ? CrossFadeState.showFirst
+                                : CrossFadeState.showSecond,
+                            duration: const Duration(milliseconds: 300),
+                          );
+                        },
+                      ),
+                    ),
+                  ],
+                ),
+              ],
+            );
+          },
+        ),
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    _controller.close();
+    super.dispose();
+  }
+}
