+++ pubspec.yaml
@@ -50,6 +50,8 @@ dependencies:
 
   syncfusion_flutter_calendar: ^20.2.36
 
+  http: ^0.13.4
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
